trytes: multiplex mod, comments, tests
