+++ trytes/src/multiplex/constants.rs
@@ -1,0 +1,8 @@
+#[cfg(target_pointer_width = "16")]
+pub const MAX_TRINARIES: usize = 16;
+#[cfg(target_pointer_width = "32")]
+pub const MAX_TRINARIES: usize = 32;
+#[cfg(target_pointer_width = "64")]
+pub const MAX_TRINARIES: usize = 64;
+#[cfg(target_pointer_width = "128")]
+pub const MAX_TRINARIES: usize = 128;
