+++ trytes/src/multiplex/mod.rs
@@ -1,0 +1,6 @@
+pub mod demux;
+pub mod constants;
+pub mod mux;
+
+pub use mux::*;
+pub use demux::*;
